Refactor the shutdown sequence of this plugin to follow the new semantic
This plugin is now interruptable, this plugin require a bit more care to
shutdown it down for a few reasons:

1. It can be stuck streaming file to disk from a S3 bucket.
2. It can be stuck reading a large file

For case 1, we cancel the streaming and delete the incomplete temp file
and not update the sincedb.

For case 2, we will cancel reading the file and not update the sincedb,
this will result in duplicates in the logging stream which is fine for now since the S3 plugin
doesn't support saving offset yet. see #54

This PR also add a fix to non deterministic test run in a local
environment by specifying a new sincedb for each run.

Fixes: #53 #42

Fixes #55

+++ lib/logstash/inputs/s3.rb
@@ -97,6 +97,7 @@ class LogStash::Inputs::S3 < LogStash::Inputs::Base
 
   public
   def run(queue)
+    @current_thread = Thread.current
     Stud.interval(@interval) do
       process_files(queue)
     end
@@ -143,15 +144,22 @@ class LogStash::Inputs::S3 < LogStash::Inputs::Base
     objects = list_new_files
 
     objects.each do |key|
-      @logger.debug("S3 input processing", :bucket => @bucket, :key => key)
-
-      lastmod = @s3bucket.objects[key].last_modified
-
-      process_log(queue, key)
-
-      sincedb.write(lastmod)
+      if stop?
+        break
+      else
+        @logger.debug("S3 input processing", :bucket => @bucket, :key => key)
+        process_log(queue, key)
+      end
     end
   end # def process_files
+
+  public
+  def stop
+    # @current_thread is initialized in the `#run` method, 
+    # this variable is needed because the `#stop` is a called in another thread 
+    # than the `#run` method and requiring us to call stop! with a explicit thread.
+    Stud.stop!(@current_thread)
+  end
 
   public
   def aws_service_endpoint(region)
@@ -164,6 +172,12 @@ class LogStash::Inputs::S3 < LogStash::Inputs::Base
   end
 
   private
+
+  # Read the content of the local file
+  #
+  # @param [Queue] Where to push the event
+  # @param [String] Which file to read from
+  # @return [Boolean] True if the file was completely read, false otherwise.
   def process_local_log(queue, filename)
     @logger.debug('Processing file', :filename => filename)
 
@@ -172,6 +186,11 @@ class LogStash::Inputs::S3 < LogStash::Inputs::Base
     # So all IO stuff: decompression, reading need to be done in the actual
     # input and send as bytes to the codecs.
     read_file(filename) do |line|
+      if stop?
+        @logger.warn("Logstash S3 input, stop reading in the middle of the file, we will read it again when logstash is started")
+        return false
+      end
+
       @codec.decode(line) do |event|
         # We are making an assumption concerning cloudfront
         # log format, the user will use the plain or the line codec
@@ -195,6 +214,8 @@ class LogStash::Inputs::S3 < LogStash::Inputs::Base
         end
       end
     end
+
+    return true
   end # def process_local_log
 
   private
@@ -296,26 +317,40 @@ class LogStash::Inputs::S3 < LogStash::Inputs::Base
     object = @s3bucket.objects[key]
 
     filename = File.join(temporary_directory, File.basename(key))
-
-    download_remote_file(object, filename)
-
-    process_local_log(queue, filename)
-
-    backup_to_bucket(object, key)
-    backup_to_dir(filename)
-
-    delete_file_from_bucket(object)
-    FileUtils.remove_entry_secure(filename, true)
+    
+    if download_remote_file(object, filename)
+      if process_local_log(queue, filename)
+        backup_to_bucket(object, key)
+        backup_to_dir(filename)
+        delete_file_from_bucket(object)
+        FileUtils.remove_entry_secure(filename, true)
+        lastmod = object.last_modified
+        sincedb.write(lastmod)
+      end
+    else
+      FileUtils.remove_entry_secure(filename, true)
+    end
   end
 
   private
+  # Stream the remove file to the local disk
+  #
+  # @param [S3Object] Reference to the remove S3 objec to download
+  # @param [String] The Temporary filename to stream to.
+  # @return [Boolean] True if the file was completely downloaded
   def download_remote_file(remote_object, local_filename)
+    completed = false
+
     @logger.debug("S3 input: Download remote file", :remote_key => remote_object.key, :local_filename => local_filename)
     File.open(local_filename, 'wb') do |s3file|
       remote_object.read do |chunk|
+        return completed if stop?
         s3file.write(chunk)
       end
     end
+    completed = true
+
+    return completed
   end
 
   private
